refactor: move celsius-fahrenheit converter to another file
